[#3605] Match the parent/child id exactly.
It was doing a simple match based on the first part of the string. This caused
multiple records to be returned which Net::LDAP would just return the last one.
This would create the wrong associated user, and would make linking to the
correct user impossible.

Caused by the changes in ed84d23e and 4a7e1963.